Add ControlPlaneRequest, ControlPlaneResponse
